Formatting. Various small fixes.

+++ src/test/java/util/command/AddOperationTest.java
@@ -32,7 +32,7 @@ public class AddOperationTest extends BaseOperationTest {
         final Operation op1 = new AddOperation("aaa", 1337);
         final Operation op2 = new AddOperation("aaa", 1337);
         Assert.assertEquals(op1, op2);
-        Assert.assertEquals(op1.hashCode(), op2.hashCode());
+        Assert.assertEquals((long) op1.hashCode(), (long) op2.hashCode());
         Assert.assertEquals(op1.toString(), op2.toString());
     }
 }
