+++ src/test/java/util/command/ReadOperationTest.java
@@ -21,7 +21,7 @@ public class ReadOperationTest extends BaseOperationTest {
         final Operation op1 = new ReadOperation("boo");
         final Operation op2 = new ReadOperation("boo");
         Assert.assertEquals(op1, op2);
-        Assert.assertEquals(op1.hashCode(), op2.hashCode());
+        Assert.assertEquals((long) op1.hashCode(), (long) op2.hashCode());
         Assert.assertEquals(op1.toString(), op2.toString());
     }
 }
